Numeric.Interval

Description

Interval arithmetic

Documentation

data Interval a

Constructors

I !a !a

(...) :: a -> a -> Interval a

The rule of thumb is you should only use this to construct using values that you took out of the interval. Otherwise, use I, to force rounding

whole :: Fractional a => Interval a

The whole real number line

empty :: Fractional a => Interval a

An empty interval

null :: Ord a => Interval a -> Bool

negation handles NaN properly

singleton :: a -> Interval a

A singleton point

elem :: Ord a => a -> Interval a -> Bool

notElem :: Ord a => a -> Interval a -> Bool

inf :: Interval a -> a

The infinumum (lower bound) of an interval

sup :: Interval a -> a

The supremum (upper bound) of an interval

singular :: Ord a => Interval a -> Bool

Is the interval a singleton point? N.B. This is fairly fragile and likely will not hold after even a few operations that only involve singletons

width :: Num a => Interval a -> a

Calculate the width of an interval.

midpoint :: Fractional a => Interval a -> a

Nearest point to the midpoint of the interval.

intersection :: (Fractional a, Ord a) => Interval a -> Interval a -> Interval a

Calculate the intersection of two intervals.

hull :: Ord a => Interval a -> Interval a -> Interval a

Calculate the convex hull of two intervals

bisection :: Fractional a => Interval a -> (Interval a, Interval a)

Bisect an interval at its midpoint.

magnitude :: (Num a, Ord a) => Interval a -> a

Magnitude

mignitude :: (Num a, Ord a) => Interval a -> a

"mignitude"

contains :: Ord a => Interval a -> Interval a -> Bool

isSubsetOf :: Ord a => Interval a -> Interval a -> Bool

certainly :: Ord a => (forall b. Ord b => b -> b -> Bool) -> Interval a -> Interval a -> Bool

For all x in X, y in Y. x op y

(<!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x < y

(<=!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x <= y

(==!) :: Eq a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x == y

(>=!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x >= y

(>!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x > y

possibly :: Ord a => (forall b. Ord b => b -> b -> Bool) -> Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x op y?

(<?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x < y?

(<=?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x <= y?

(==?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x == y?

(>=?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x >= y?

(>?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x > y?

idouble :: Interval Double -> Interval Double

ifloat :: Interval Float -> Interval Float