Numeric.Interval.NonEmpty

Description

Interval arithmetic

Documentation

data Interval a

(...) :: Ord a => a -> a -> Interval a infix 3

Create a non-empty interval, turning it around if necessary

interval :: Ord a => a -> a -> Maybe (Interval a)

Try to create a non-empty interval.

whole :: Fractional a => Interval a

The whole real number line

>>> whole
-Infinity ... Infinity

singleton :: a -> Interval a

A singleton point

>>> singleton 1
1 ... 1

elem :: Ord a => a -> Interval a -> Bool

Determine if a point is in the interval.

>>> elem 3.2 (1.0 ... 5.0)
True

>>> elem 5 (1.0 ... 5.0)
True

>>> elem 1 (1.0 ... 5.0)
True

>>> elem 8 (1.0 ... 5.0)
False

notElem :: Ord a => a -> Interval a -> Bool

Determine if a point is not included in the interval

>>> notElem 8 (1.0 ... 5.0)
True

>>> notElem 1.4 (1.0 ... 5.0)
False

inf :: Interval a -> a

The infinumum (lower bound) of an interval

>>> inf (1 ... 20)
1

sup :: Interval a -> a

The supremum (upper bound) of an interval

>>> sup (1 ... 20)
20

singular :: Ord a => Interval a -> Bool

Is the interval a singleton point? N.B. This is fairly fragile and likely will not hold after even a few operations that only involve singletons

>>> singular (singleton 1)
True

>>> singular (1.0 ... 20.0)
False

width :: Num a => Interval a -> a

Calculate the width of an interval.

>>> width (1 ... 20)
19

>>> width (singleton 1)
0

midpoint :: Fractional a => Interval a -> a

Nearest point to the midpoint of the interval.

>>> midpoint (10.0 ... 20.0)
15.0

>>> midpoint (singleton 5.0)
5.0

intersection :: (Fractional a, Ord a) => Interval a -> Interval a -> Maybe (Interval a)

Calculate the intersection of two intervals.

>>> intersection (1 ... 10 :: Interval Double) (5 ... 15 :: Interval Double)
Just (5.0 ... 10.0)

hull :: Ord a => Interval a -> Interval a -> Interval a

Calculate the convex hull of two intervals

>>> hull (0 ... 10 :: Interval Double) (5 ... 15 :: Interval Double)
0.0 ... 15.0

>>> hull (15 ... 85 :: Interval Double) (0 ... 10 :: Interval Double)
0.0 ... 85.0

bisect :: Fractional a => Interval a -> (Interval a, Interval a)

Bisect an interval at its midpoint.

>>> bisect (10.0 ... 20.0)
(10.0 ... 15.0,15.0 ... 20.0)

>>> bisect (singleton 5.0)
(5.0 ... 5.0,5.0 ... 5.0)

bisectIntegral :: Integral a => Interval a -> (Interval a, Interval a)

magnitude :: (Num a, Ord a) => Interval a -> a

Magnitude

>>> magnitude (1 ... 20)
20

>>> magnitude (-20 ... 10)
20

>>> magnitude (singleton 5)
5

mignitude :: (Num a, Ord a) => Interval a -> a

"mignitude"

>>> mignitude (1 ... 20)
1

>>> mignitude (-20 ... 10)
0

>>> mignitude (singleton 5)
5

contains :: Ord a => Interval a -> Interval a -> Bool

Check if interval X totally contains interval Y

>>> (20 ... 40 :: Interval Double) `contains` (25 ... 35 :: Interval Double)
True

>>> (20 ... 40 :: Interval Double) `contains` (15 ... 35 :: Interval Double)
False

isSubsetOf :: Ord a => Interval a -> Interval a -> Bool

Flipped version of contains. Check if interval X a subset of interval Y

>>> (25 ... 35 :: Interval Double) `isSubsetOf` (20 ... 40 :: Interval Double)
True

>>> (20 ... 40 :: Interval Double) `isSubsetOf` (15 ... 35 :: Interval Double)
False

certainly :: Ord a => (forall b. Ord b => b -> b -> Bool) -> Interval a -> Interval a -> Bool

For all x in X, y in Y. x op y

(<!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x < y

>>> (5 ... 10 :: Interval Double) <! (20 ... 30 :: Interval Double)
True

>>> (5 ... 10 :: Interval Double) <! (10 ... 30 :: Interval Double)
False

>>> (20 ... 30 :: Interval Double) <! (5 ... 10 :: Interval Double)
False

(<=!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x <= y

>>> (5 ... 10 :: Interval Double) <=! (20 ... 30 :: Interval Double)
True

>>> (5 ... 10 :: Interval Double) <=! (10 ... 30 :: Interval Double)
True

>>> (20 ... 30 :: Interval Double) <=! (5 ... 10 :: Interval Double)
False

(==!) :: Eq a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x == y

Only singleton intervals or empty intervals can return true

>>> (singleton 5 :: Interval Double) ==! (singleton 5 :: Interval Double)
True

>>> (5 ... 10 :: Interval Double) ==! (5 ... 10 :: Interval Double)
False

(>=!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x >= y

>>> (20 ... 40 :: Interval Double) >=! (10 ... 20 :: Interval Double)
True

>>> (5 ... 20 :: Interval Double) >=! (15 ... 40 :: Interval Double)
False

(>!) :: Ord a => Interval a -> Interval a -> Bool

For all x in X, y in Y. x > y

>>> (20 ... 40 :: Interval Double) >! (10 ... 19 :: Interval Double)
True

>>> (5 ... 20 :: Interval Double) >! (15 ... 40 :: Interval Double)
False

possibly :: Ord a => (forall b. Ord b => b -> b -> Bool) -> Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x op y?

(<?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x < y?

(<=?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x <= y?

(==?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x == y?

(>=?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x >= y?

(>?) :: Ord a => Interval a -> Interval a -> Bool

Does there exist an x in X, y in Y such that x > y?

clamp :: Ord a => Interval a -> a -> a

The nearest value to that supplied which is contained in the interval.

idouble :: Interval Double -> Interval Double

id function. Useful for type specification

>>> :t idouble (1 ... 3)
idouble (1 ... 3) :: Interval Double

ifloat :: Interval Float -> Interval Float

id function. Useful for type specification

>>> :t ifloat (1 ... 3)
ifloat (1 ... 3) :: Interval Float